Sky Rapor ve FormBuilder düzenlemeler

+++ ortak/formBuilder/formBuilderBase.js
@@ -75,7 +75,8 @@ class FormBuilderBase extends CObject {
 		$.extend(this, {
 			_id: e.id, _parent: e.parent, _layout: e.layout, _part: e.part, _parentPart: e.parentPart, _visibleKosulu: e.visibleKosulu == null ? e.visibleBlock : e.visibleKosulu,
 			autoAppendFlag: e.autoAppend ?? e.autoAppendFlag ?? this.defaultAutoAppendFlag, autoAppendIslemi: e.autoAppendIslemi, autoInitLayoutFlag: e.autoInitLayout ?? e.autoInitLayoutFlag,
-			builders: e.builders || [], args: e.args, init: e.init, buildEk: e.buildEk, afterRun: e.afterRun, id2Builder: null, styles: e.styles || [], cssClasses: e.cssClasses || [], userData: e.userData
+			builders: e.builders || [], args: e.args, init: e.init, buildEk: e.buildEk, afterRun: e.afterRun, id2Builder: null, styles: e.styles || [], cssClasses: e.cssClasses || [],
+			canDestroy: e.canDestroy ?? true, userData: e.userData
 		});
 		if (!this.autoAppendIslemi) { this.autoAppendMode_append() }
 		const {sabitDegerler} = e; if (sabitDegerler) { for (const key in sabitDegerler) { const value = sabitDegerler[key]; if (value !== undefined) { this[key] = value } } }
@@ -214,9 +215,11 @@ class FormBuilderBase extends CObject {
 	destroyPart(e) {
 		if (this.isDestroyed || this._destroying) { return } this._destroying = true;
 		const rootBuilder = this.rootBuilder ?? this, {builders, id} = this, {CSSClass_BuilderId} = FormBuilderBase;
-		try { const styles = $(`style[data-${CSSClass_BuilderId} = "${id}"]`); if (styles?.length) { styles.remove() } } catch (ex) { }
-		try { const {layout, input, part} = this; if (part) { if (part.destroyPart) { part.destroyPart(e) } } for (const elm of [input, layout]) { if (elm?.length) { elm.remove()} } } catch (ex) { console.error(ex) }
-		try { for (const builder of builders) { if (builder != this && builder.destroyPart) { builder.destroyPart(e) } } } catch (ex) { console.error(ex) }
+		if (this.canDestroy) {
+			try { const styles = $(`style[data-${CSSClass_BuilderId} = "${id}"]`); if (styles?.length) { styles.remove() } } catch (ex) { }
+			try { const {layout, input, part} = this; if (part) { if (part.destroyPart) { part.destroyPart(e) } } for (const elm of [input, layout]) { if (elm?.length) { elm.remove()} } } catch (ex) { console.error(ex) }
+			try { for (const builder of builders) { if (builder != this && builder.destroyPart) { builder.destroyPart(e) } } } catch (ex) { console.error(ex) }
+		}
 		delete this._destroying; this.isDestroyed = true
 		/*if (!rootBuilder.isDestroyed) { const {CSSClass_BuilderId} = FormBuilderBase, styles = $(`style[data-${CSSClass_BuilderId}="${this.id}"]`); if (styles?.length) { styles.remove() } rootBuilder.isDestroyed = true }*/
 	}
@@ -608,6 +611,8 @@ class FormBuilderBase extends CObject {
 	setParentPart(value) { this.parentPart = value; return this }
 	setParent(value) { this.parent = value; return this }
 	setLayout(value) { this.layout = value; return this }
+	destroyable() { this.canDestroy = true; return this }
+	noDestroy() { this.canDestroy = false; return this }
 	addStyle_wh(e, _height) {
 		if (e == null || typeof e != 'object') e = { width: e, height: _height }
 		const withImportant = value => {
